feat(web-ui): surface scoped deep review reports
