+++ src/web-ui/src/flow_chat/deep-review/action-bar/DeepReviewActionBar.tsx
@@ -10,20 +10,16 @@ import {
   ChevronUp,
   MessageSquare,
   Play,
-  Pause,
   Copy,
   Info,
   SkipForward,
   RotateCcw,
   Eye,
-  Minus,
-  Settings,
 } from 'lucide-react';
 import { Button, Checkbox, Tooltip } from '@/component-library';
 import {
   useReviewActionBarStore,
   type DeepReviewCapacityQueueAction,
-  type DeepReviewCapacityQueueReason,
   type ReviewActionPhase,
 } from '../../store/deepReviewActionBarStore';
 import type { ReviewRemediationItem } from '../../utils/codeReviewRemediation';
@@ -50,12 +46,14 @@ import {
   extractPartialReviewData,
 } from '../../utils/deepReviewExperience';
 import { flowChatStore } from '../../store/FlowChatStore';
-import { CodeReviewReportExportActions } from '../../tool-cards/CodeReviewReportExportActions';
 import { agentAPI } from '@/infrastructure/api/service-api/AgentAPI';
 import { lowerDefaultReviewTeamMaxParallelReviewers } from '@/shared/services/reviewTeamService';
 import { useSettingsStore } from '@/app/scenes/settings/settingsStore';
 import { useSceneStore } from '@/app/stores/sceneStore';
 import type { ConfigTab } from '@/app/scenes/settings/settingsConfig';
+import { formatElapsedTime } from './actionBarFormatting';
+import { CapacityQueueNotice } from './CapacityQueueNotice';
+import { ReviewActionHeader } from './ReviewActionHeader';
 import '../../components/btw/DeepReviewActionBar.scss';
 
 const log = createLogger('DeepReviewActionBar');
@@ -90,14 +88,6 @@ const GROUP_PRIORITY_META: Record<RemediationGroupId, { color: string }> = {
   should_improve: { color: 'var(--color-warning, #f59e0b)' },
   needs_decision: { color: 'var(--color-accent-500, #60a5fa)' },
   verification: { color: 'var(--color-success, #22c55e)' },
-};
-
-const CAPACITY_QUEUE_REASON_KEYS: Record<DeepReviewCapacityQueueReason, string> = {
-  provider_rate_limit: 'deepReviewActionBar.capacityQueue.reasons.providerRateLimit',
-  provider_concurrency_limit: 'deepReviewActionBar.capacityQueue.reasons.providerConcurrencyLimit',
-  retry_after: 'deepReviewActionBar.capacityQueue.reasons.retryAfter',
-  local_concurrency_cap: 'deepReviewActionBar.capacityQueue.reasons.localConcurrencyCap',
-  temporary_overload: 'deepReviewActionBar.capacityQueue.reasons.temporaryOverload',
 };
 
 const stopNestedScrollPropagation = (event: React.WheelEvent | React.TouchEvent) => {
@@ -156,18 +146,6 @@ export const ReviewActionBar: React.FC = () => {
     capacityQueueState?.controlMode === 'backend'
       ? hasBackendQueueControlTarget
       : capacityQueueState?.controlMode !== 'session_stop_only';
-  const capacityQueueReasonLabel = capacityQueueState?.reason
-    ? t(CAPACITY_QUEUE_REASON_KEYS[capacityQueueState.reason], {
-      defaultValue: capacityQueueState.reason.split('_').join(' '),
-    })
-    : null;
-  const capacityQueueElapsedLabel = capacityQueueState?.queueElapsedMs !== undefined
-    ? formatElapsedTime(capacityQueueState.queueElapsedMs)
-    : null;
-  const capacityQueueMaxWaitLabel = capacityQueueState?.maxQueueWaitSeconds !== undefined
-    ? formatElapsedTime(capacityQueueState.maxQueueWaitSeconds * 1000)
-    : null;
-
   const handleCapacityQueueAction = useCallback(async (
     action: DeepReviewCapacityQueueAction,
     applyLocalAction: () => void,
@@ -718,33 +696,15 @@ export const ReviewActionBar: React.FC = () => {
       onWheel={stopNestedScrollPropagation}
       onTouchMove={stopNestedScrollPropagation}
     >
-      {/* Top-right controls: export actions + minimize */}
-      <div className="deep-review-action-bar__controls">
-        {reviewData && (
-          <CodeReviewReportExportActions reviewData={reviewData} />
-        )}
-        <span className="deep-review-action-bar__controls-divider" />
-        <button
-          type="button"
-          className="deep-review-action-bar__controls-btn"
-          onClick={handleMinimize}
-          aria-label={t('deepReviewActionBar.minimize', { defaultValue: 'Minimize' })}
-        >
-          <Minus size={14} />
-        </button>
-      </div>
-
-      {/* Phase status header */}
-      <div className="deep-review-action-bar__status">
-        <PhaseIcon
-          size={18}
-          className={`deep-review-action-bar__icon ${phaseConfig.iconClass}`}
-        />
-        <span className="deep-review-action-bar__status-title">{phaseTitle}</span>
-        {errorMessage && (
-          <span className="deep-review-action-bar__error-message">{errorMessage}</span>
-        )}
-      </div>
+      <ReviewActionHeader
+        reviewData={reviewData}
+        PhaseIcon={PhaseIcon}
+        phaseIconClass={phaseConfig.iconClass}
+        phaseTitle={phaseTitle}
+        errorMessage={errorMessage}
+        minimizeLabel={t('deepReviewActionBar.minimize', { defaultValue: 'Minimize' })}
+        onMinimize={handleMinimize}
+      />
 
       {/* Running progress */}
       {(phase === 'fix_running' || phase === 'resume_running') && progressSummary && (
@@ -765,148 +725,28 @@ export const ReviewActionBar: React.FC = () => {
 
       {/* Capacity queue notice */}
       {showCapacityQueueNotice && capacityQueueState && (
-        <div className="deep-review-action-bar__capacity-queue" aria-live="polite">
-          <div className="deep-review-action-bar__capacity-queue-main">
-            <Clock size={14} className="deep-review-action-bar__capacity-queue-icon" />
-            <div className="deep-review-action-bar__capacity-queue-copy">
-              <span className="deep-review-action-bar__capacity-queue-title">
-                {capacityQueueState.status === 'paused_by_user'
-                  ? t('deepReviewActionBar.capacityQueue.pausedTitle', {
-                    defaultValue: 'Queue paused',
-                  })
-                  : t('deepReviewActionBar.capacityQueue.title', {
-                    defaultValue: 'Reviewers waiting for capacity',
-                  })}
-              </span>
-              <span className="deep-review-action-bar__capacity-queue-detail">
-                {t('deepReviewActionBar.capacityQueue.detail', {
-                  defaultValue: 'Queue wait does not count against reviewer runtime.',
-                })}
-              </span>
-              {(capacityQueueReasonLabel || capacityQueueElapsedLabel) && (
-                <span className="deep-review-action-bar__capacity-queue-meta">
-                  {capacityQueueReasonLabel && (
-                    <span className="deep-review-action-bar__capacity-queue-chip">
-                      {t('deepReviewActionBar.capacityQueue.reason', {
-                        reason: capacityQueueReasonLabel,
-                        defaultValue: `Reason: ${capacityQueueReasonLabel}`,
-                      })}
-                    </span>
-                  )}
-                  {capacityQueueElapsedLabel && (
-                    <span className="deep-review-action-bar__capacity-queue-chip">
-                      {capacityQueueMaxWaitLabel
-                        ? t('deepReviewActionBar.capacityQueue.elapsedWithMax', {
-                          elapsed: capacityQueueElapsedLabel,
-                          max: capacityQueueMaxWaitLabel,
-                          defaultValue: `Waited ${capacityQueueElapsedLabel} of ${capacityQueueMaxWaitLabel}`,
-                        })
-                        : t('deepReviewActionBar.capacityQueue.elapsed', {
-                          elapsed: capacityQueueElapsedLabel,
-                          defaultValue: `Waited ${capacityQueueElapsedLabel}`,
-                        })}
-                    </span>
-                  )}
-                </span>
-              )}
-              {capacityQueueState.sessionConcurrencyHigh && (
-                <span className="deep-review-action-bar__capacity-queue-detail">
-                  {t('deepReviewActionBar.capacityQueue.sessionBusy', {
-                    defaultValue: 'Your active session is busy. Pause Deep Review or continue later.',
-                  })}
-                </span>
-              )}
-              {!supportsInlineQueueControls && (
-                <span className="deep-review-action-bar__capacity-queue-detail">
-                  {t('deepReviewActionBar.capacityQueue.stopHint', {
-                    defaultValue: 'Use Stop to interrupt this review queue.',
-                  })}
-                </span>
-              )}
-            </div>
-          </div>
-          <div className="deep-review-action-bar__capacity-queue-actions">
-            {supportsInlineQueueControls && (
-              <>
-                {capacityQueueState.status === 'paused_by_user' ? (
-                  <Button
-                    variant="secondary"
-                    size="small"
-                    onClick={() => void handleCapacityQueueAction(
-                      'continue',
-                      store.continueCapacityQueue,
-                    )}
-                  >
-                    <Play size={13} />
-                    {t('deepReviewActionBar.capacityQueue.continueQueue', {
-                      defaultValue: 'Continue queue',
-                    })}
-                  </Button>
-                ) : (
-                  <Button
-                    variant="secondary"
-                    size="small"
-                    onClick={() => void handleCapacityQueueAction(
-                      'pause',
-                      store.pauseCapacityQueue,
-                    )}
-                  >
-                    <Pause size={13} />
-                    {t('deepReviewActionBar.capacityQueue.pauseQueue', {
-                      defaultValue: 'Pause queue',
-                    })}
-                  </Button>
-                )}
-                {(capacityQueueState.optionalReviewerCount ?? 0) > 0 && (
-                  <Button
-                    variant="ghost"
-                    size="small"
-                    onClick={() => void handleCapacityQueueAction(
-                      'skip_optional',
-                      store.skipOptionalQueuedReviewers,
-                    )}
-                  >
-                    <SkipForward size={13} />
-                    {t('deepReviewActionBar.capacityQueue.skipOptionalQueued', {
-                      defaultValue: 'Skip optional extras',
-                    })}
-                  </Button>
-                )}
-                <Button
-                  variant="ghost"
-                  size="small"
-                  onClick={() => void handleCapacityQueueAction(
-                    'cancel',
-                    store.cancelQueuedReviewers,
-                  )}
-                >
-                  {t('deepReviewActionBar.capacityQueue.cancelQueued', {
-                    defaultValue: 'Cancel queued reviewers',
-                  })}
-                </Button>
-              </>
-            )}
-            <Button
-              variant="ghost"
-              size="small"
-              onClick={() => void handleRunSlowerNextTime()}
-            >
-              <Settings size={13} />
-              {t('deepReviewActionBar.capacityQueue.runSlowerNextTime', {
-                defaultValue: 'Run slower next time',
-              })}
-            </Button>
-            <Button
-              variant="ghost"
-              size="small"
-              onClick={handleOpenReviewSettings}
-            >
-              {t('deepReviewActionBar.capacityQueue.openReviewSettings', {
-                defaultValue: 'Open Review settings',
-              })}
-            </Button>
-          </div>
-        </div>
+        <CapacityQueueNotice
+          capacityQueueState={capacityQueueState}
+          supportsInlineQueueControls={supportsInlineQueueControls}
+          onContinueQueue={() => handleCapacityQueueAction(
+            'continue',
+            store.continueCapacityQueue,
+          )}
+          onPauseQueue={() => handleCapacityQueueAction(
+            'pause',
+            store.pauseCapacityQueue,
+          )}
+          onSkipOptionalQueuedReviewers={() => handleCapacityQueueAction(
+            'skip_optional',
+            store.skipOptionalQueuedReviewers,
+          )}
+          onCancelQueuedReviewers={() => handleCapacityQueueAction(
+            'cancel',
+            store.cancelQueuedReviewers,
+          )}
+          onRunSlowerNextTime={handleRunSlowerNextTime}
+          onOpenReviewSettings={handleOpenReviewSettings}
+        />
       )}
 
       {/* Partial results summary on interruption */}
@@ -1452,15 +1292,5 @@ export const ReviewActionBar: React.FC = () => {
     </div>
   );
 };
-
-function formatElapsedTime(ms: number): string {
-  const seconds = Math.floor(ms / 1000);
-  const minutes = Math.floor(seconds / 60);
-  const remainingSeconds = seconds % 60;
-  if (minutes > 0) {
-    return `${minutes}m ${remainingSeconds}s`;
-  }
-  return `${seconds}s`;
-}
 
 export const DeepReviewActionBar = ReviewActionBar;
