+++ src/web-ui/src/flow_chat/deep-review/action-bar/ReviewActionHeader.test.tsx
@@ -1,0 +1,30 @@
+import React from 'react';
+import { renderToStaticMarkup } from 'react-dom/server';
+import { describe, expect, it, vi } from 'vitest';
+import { ReviewActionHeader } from './ReviewActionHeader';
+
+vi.mock('../../tool-cards/CodeReviewReportExportActions', () => ({
+  CodeReviewReportExportActions: () => <span>export actions</span>,
+}));
+
+describe('ReviewActionHeader', () => {
+  it('renders export actions, status, error, and minimize control', () => {
+    const Icon = () => <span>phase icon</span>;
+    const html = renderToStaticMarkup(
+      <ReviewActionHeader
+        reviewData={{ summary: { recommended_action: 'request_changes' } } as any}
+        PhaseIcon={Icon}
+        phaseIconClass="phase-class"
+        phaseTitle="Review completed"
+        errorMessage="Network warning"
+        minimizeLabel="Minimize"
+        onMinimize={vi.fn()}
+      />,
+    );
+
+    expect(html).toContain('export actions');
+    expect(html).toContain('Review completed');
+    expect(html).toContain('Network warning');
+    expect(html).toContain('aria-label="Minimize"');
+  });
+});
