+++ src/web-ui/src/locales/en-US/flow-chat.json
@@ -749,6 +749,12 @@
     "estimatedTokens": "Estimated: {{min}} - {{max}} tokens",
     "runStrategy": "Run strategy: {{strategy}}",
     "recommendedStrategy": "Recommended strategy: {{strategy}}",
+    "reviewDepth": "Review depth: {{depth}}",
+    "reviewDepthLabels": {
+      "high_risk_only": "High-risk-only",
+      "risk_expanded": "Risk-expanded",
+      "full_depth": "Full-depth"
+    },
     "recommendationTitle": "Risk recommendation",
     "strategyOverrideTitle": "Run strategy",
     "strategyOverrideBody": "Choose a project-specific strategy for this launch.",
@@ -1514,7 +1520,13 @@
       },
       "runManifest": {
         "recommendedStrategy": "Recommended strategy",
-        "riskRecommendationTitle": "Risk recommendation"
+        "riskRecommendationTitle": "Risk recommendation",
+        "reviewDepth": "Review depth",
+        "reviewDepthLabels": {
+          "high_risk_only": "High-risk-only",
+          "risk_expanded": "Risk-expanded",
+          "full_depth": "Full-depth"
+        }
       },
       "sectionItemCount": "{{count}} items",
       "remediationPlan": "Remediation Plan",
@@ -1543,6 +1555,10 @@
         "partial_reviewer": {
           "label": "Reviewer timed out with partial result",
           "detail": "{{count}} reviewer result is partial; confidence is reduced."
+        },
+        "reduced_scope": {
+          "label": "Reduced-depth coverage",
+          "detail": "This review used a reduced-depth scope profile."
         },
         "retry_guidance": {
           "label": "Retry guidance emitted",
