+++ src/web-ui/src/shared/services/review-team/index.ts
@@ -27,68 +27,65 @@ import {
   DISALLOWED_REVIEW_TEAM_MEMBER_IDS,
   EXTRA_MEMBER_DEFAULTS,
   FALLBACK_REVIEW_TEAM_DEFINITION,
-  JUDGE_WORK_PACKET_REQUIRED_OUTPUT_FIELDS,
   MAX_AUTO_RETRY_ELAPSED_GUARD_SECONDS,
   MAX_PARALLEL_REVIEWER_INSTANCES,
-  MAX_PREDICTIVE_TIMEOUT_SECONDS,
   MAX_QUEUE_WAIT_SECONDS,
-  PREDICTIVE_TIMEOUT_BASE_SECONDS,
-  PREDICTIVE_TIMEOUT_PER_100_LINES_SECONDS,
-  PREDICTIVE_TIMEOUT_PER_FILE_SECONDS,
-  PROMPT_BYTE_ESTIMATE_BASE_BYTES,
-  PROMPT_BYTE_ESTIMATE_PER_CHANGED_LINE_BYTES,
-  PROMPT_BYTE_ESTIMATE_PER_FILE_BYTES,
-  PROMPT_BYTE_ESTIMATE_UNKNOWN_LINES_PER_FILE,
-  REVIEWER_WORK_PACKET_REQUIRED_OUTPUT_FIELDS,
   REVIEW_WORK_PACKET_ALLOWED_TOOLS,
-  TOKEN_BUDGET_PROMPT_BYTE_LIMIT_BY_MODE,
 } from './defaults';
 import {
-  REVIEW_STRATEGY_COMMON_RULES,
   REVIEW_STRATEGY_LEVELS,
   REVIEW_STRATEGY_PROFILES,
-  getReviewStrategyProfile,
 } from './strategy';
+import { buildPreReviewSummary } from './preReviewSummary';
+import {
+  buildIncrementalReviewCachePlan,
+  buildSharedContextCachePlan,
+} from './cachePlan';
+import { buildDeepReviewEvidencePack } from './evidencePack';
+import {
+  applyTeamStrategyOverrideToMember,
+  toManifestMember,
+} from './manifestMembers';
+import {
+  buildReviewStrategyDecision,
+  recommendBackendCompatibleStrategyForTarget,
+  recommendReviewStrategyForTarget,
+} from './risk';
+import { buildDeepReviewScopeProfile } from './scopeProfile';
+import {
+  buildEffectiveExecutionPolicy,
+  buildTokenBudgetPlan,
+} from './tokenBudget';
+import {
+  buildWorkPackets,
+  resolveChangeStats,
+  resolveMaxExtraReviewers,
+} from './workPackets';
+import { buildReviewTeamPromptBlockContent } from './promptBlock';
 import type {
   ReviewMemberStrategyLevel,
   ReviewModelFallbackReason,
-  ReviewRoleDirectiveKey,
   ReviewStrategyLevel,
   ReviewStrategyProfile,
   ReviewStrategySource,
   ReviewTeam,
-  ReviewTeamBackendRiskFactors,
-  ReviewTeamBackendStrategyRecommendation,
   ReviewTeamChangeStats,
   ReviewTeamConcurrencyPolicy,
   ReviewTeamCoreRoleDefinition,
   ReviewTeamCoreRoleKey,
   ReviewTeamDefinition,
   ReviewTeamExecutionPolicy,
-  ReviewTeamIncrementalReviewCacheInvalidation,
-  ReviewTeamIncrementalReviewCachePlan,
-  ReviewTeamManifestMember,
   ReviewTeamManifestMemberReason,
   ReviewTeamMember,
-  ReviewTeamPreReviewSummary,
   ReviewTeamRateLimitStatus,
-  ReviewTeamRiskFactors,
   ReviewTeamRunManifest,
-  ReviewTeamSharedContextCachePlan,
-  ReviewTeamSharedContextTool,
   ReviewTeamStoredConfig,
-  ReviewTeamStrategyDecision,
-  ReviewTeamStrategyMismatchSeverity,
-  ReviewTeamStrategyRecommendation,
-  ReviewTeamTokenBudgetDecision,
-  ReviewTeamTokenBudgetPlan,
-  ReviewTeamWorkPacket,
-  ReviewTeamWorkPacketScope,
   ReviewTokenBudgetMode,
 } from './types';
 
 export * from './types';
 export * from './strategy';
+export { recommendReviewStrategyForTarget } from './risk';
 export {
   DEFAULT_REVIEW_TEAM_ID,
   DEFAULT_REVIEW_TEAM_CONFIG_PATH,
@@ -1235,1032 +1232,11 @@ export async function prepareDefaultReviewTeamForLaunch(
   return team;
 }
 
-function toManifestMember(
-  member: ReviewTeamMember,
-  reason?: ReviewTeamManifestMember['reason'],
-): ReviewTeamManifestMember {
-  const strategyProfile = getReviewStrategyProfile(member.strategyLevel);
-  const roleDirective =
-    strategyProfile.roleDirectives[member.subagentId as ReviewRoleDirectiveKey];
-  return {
-    subagentId: member.subagentId,
-    displayName: member.displayName,
-    roleName: member.roleName,
-    model: member.model || DEFAULT_REVIEW_TEAM_MODEL,
-    configuredModel: member.configuredModel || member.model || DEFAULT_REVIEW_TEAM_MODEL,
-    modelFallbackReason: member.modelFallbackReason,
-    defaultModelSlot: member.defaultModelSlot ?? strategyProfile.defaultModelSlot,
-    strategyLevel: member.strategyLevel,
-    strategySource: member.strategySource,
-    strategyDirective:
-      member.strategyDirective || roleDirective || strategyProfile.promptDirective,
-    locked: member.locked,
-    source: member.source,
-    subagentSource: member.subagentSource,
-    ...(reason ? { reason } : {}),
-  };
-}
-
-function resolveManifestMemberModelForStrategy(
-  member: ReviewTeamMember,
-  strategyLevel: ReviewStrategyLevel,
-): {
-  model: string;
-  configuredModel: string;
-  modelFallbackReason?: ReviewModelFallbackReason;
-} {
-  if (member.modelFallbackReason === 'model_removed') {
-    return {
-      model: getReviewStrategyProfile(strategyLevel).defaultModelSlot,
-      configuredModel: member.configuredModel,
-      modelFallbackReason: member.modelFallbackReason,
-    };
-  }
-
-  return resolveMemberModel(
-    member.configuredModel || member.model || DEFAULT_REVIEW_TEAM_MODEL,
-    strategyLevel,
-  );
-}
-
-function applyTeamStrategyOverrideToMember(
-  member: ReviewTeamMember,
-  strategyLevel: ReviewStrategyLevel,
-): ReviewTeamMember {
-  if (member.strategySource === 'member' || member.strategyLevel === strategyLevel) {
-    return member;
-  }
-
-  const strategyProfile = getReviewStrategyProfile(strategyLevel);
-  const model = resolveManifestMemberModelForStrategy(member, strategyLevel);
-  return {
-    ...member,
-    model: model.model,
-    configuredModel: model.configuredModel,
-    modelFallbackReason: model.modelFallbackReason,
-    strategyOverride: DEFAULT_REVIEW_MEMBER_STRATEGY_LEVEL,
-    strategyLevel,
-    strategySource: 'team',
-    defaultModelSlot: strategyProfile.defaultModelSlot,
-    strategyDirective:
-      strategyProfile.roleDirectives[member.subagentId as ReviewRoleDirectiveKey] ||
-      strategyProfile.promptDirective,
-  };
-}
-
 function shouldRunCoreReviewerForTarget(
   member: ReviewTeamMember,
   target: ReviewTargetClassification,
 ): boolean {
   return shouldRunReviewerForTarget(member.subagentId, target);
-}
-
-function resolveMaxExtraReviewers(
-  mode: ReviewTokenBudgetMode,
-  eligibleExtraReviewerCount: number,
-): number {
-  if (mode === 'economy') {
-    return 0;
-  }
-  return eligibleExtraReviewerCount;
-}
-
-function resolveChangeStats(
-  target: ReviewTargetClassification,
-  stats?: Partial<ReviewTeamChangeStats>,
-): ReviewTeamChangeStats {
-  const fileCount = Math.max(
-    0,
-    Math.floor(
-      stats?.fileCount ??
-        target.files.filter((file) => !file.excluded).length,
-    ),
-  );
-  const totalLinesChanged =
-    typeof stats?.totalLinesChanged === 'number' &&
-    Number.isFinite(stats.totalLinesChanged)
-      ? Math.max(0, Math.floor(stats.totalLinesChanged))
-      : undefined;
-
-  return {
-    fileCount,
-    ...(totalLinesChanged !== undefined ? { totalLinesChanged } : {}),
-    lineCountSource:
-      totalLinesChanged !== undefined
-        ? stats?.lineCountSource ?? 'diff_stat'
-        : 'unknown',
-  };
-}
-
-const SECURITY_SENSITIVE_PATH_PATTERN =
-  /(^|[/._-])(auth|oauth|crypto|security|permission|permissions|secret|secrets|token|tokens|credential|credentials)([/._-]|$)/;
-
-function isSecuritySensitiveReviewPath(normalizedPath: string): boolean {
-  return SECURITY_SENSITIVE_PATH_PATTERN.test(normalizedPath.toLowerCase());
-}
-
-function workspaceAreaForReviewPath(normalizedPath: string): string {
-  const crateMatch = normalizedPath.match(/^src\/crates\/([^/]+)/);
-  if (crateMatch) {
-    return `crate:${crateMatch[1]}`;
-  }
-
-  const appMatch = normalizedPath.match(/^src\/apps\/([^/]+)/);
-  if (appMatch) {
-    return `app:${appMatch[1]}`;
-  }
-
-  if (normalizedPath.startsWith('src/web-ui/')) {
-    return 'web-ui';
-  }
-
-  if (normalizedPath.startsWith('BitFun-Installer/')) {
-    return 'installer';
-  }
-
-  const [root] = normalizedPath.split('/');
-  return root || 'unknown';
-}
-
-function pluralize(count: number, singular: string): string {
-  return `${count} ${singular}${count === 1 ? '' : 's'}`;
-}
-
-const PRE_REVIEW_SUMMARY_SAMPLE_FILE_LIMIT = 3;
-const PRE_REVIEW_SUMMARY_AREA_LIMIT = 8;
-
-function buildPreReviewSummary(
-  target: ReviewTargetClassification,
-  changeStats: ReviewTeamChangeStats,
-): ReviewTeamPreReviewSummary {
-  const includedFiles = target.files
-    .filter((file) => !file.excluded)
-    .map((file) => file.normalizedPath);
-  const excludedFileCount = target.files.length - includedFiles.length;
-  const allWorkspaceAreas = groupFilesByWorkspaceArea(includedFiles)
-    .sort((a, b) => b.files.length - a.files.length || a.index - b.index);
-  const workspaceAreas = allWorkspaceAreas
-    .slice(0, PRE_REVIEW_SUMMARY_AREA_LIMIT)
-    .map((area) => ({
-      key: area.key,
-      fileCount: area.files.length,
-      sampleFiles: area.files.slice(0, PRE_REVIEW_SUMMARY_SAMPLE_FILE_LIMIT),
-    }));
-  const lineCount = changeStats.totalLinesChanged;
-  const lineCountLabel =
-    lineCount === undefined
-      ? 'unknown changed lines'
-      : `${lineCount} changed lines`;
-  const areaLabel = workspaceAreas.length > 0
-    ? workspaceAreas.map((area) => `${area.key} (${area.fileCount})`).join(', ')
-    : 'no resolved workspace area';
-  const targetTags = [...target.tags];
-  const tagLabel = targetTags.filter((tag) => tag !== 'unknown').join(', ') || 'unknown';
-  const omittedAreaCount = Math.max(
-    0,
-    allWorkspaceAreas.length - workspaceAreas.length,
-  );
-  const summaryParts = [
-    `${pluralize(changeStats.fileCount, 'file')}, ${lineCountLabel} across ${pluralize(allWorkspaceAreas.length, 'workspace area')}: ${areaLabel}`,
-    `tags: ${tagLabel}`,
-    omittedAreaCount > 0 ? `${pluralize(omittedAreaCount, 'workspace area')} omitted from summary` : undefined,
-  ].filter(Boolean);
-
-  return {
-    source: 'target_manifest',
-    summary: summaryParts.join('; '),
-    fileCount: changeStats.fileCount,
-    excludedFileCount,
-    ...(lineCount !== undefined ? { lineCount } : {}),
-    lineCountSource: changeStats.lineCountSource,
-    targetTags,
-    workspaceAreas,
-    warnings: target.warnings.map((warning) => warning.code),
-  };
-}
-
-export function recommendReviewStrategyForTarget(
-  target: ReviewTargetClassification,
-  changeStats: ReviewTeamChangeStats,
-): ReviewTeamStrategyRecommendation {
-  const includedFiles = target.files.filter((file) => !file.excluded);
-  const securityFileCount = includedFiles.filter((file) =>
-    isSecuritySensitiveReviewPath(file.normalizedPath),
-  ).length;
-  const workspaceAreaCount = new Set(
-    includedFiles.map((file) => workspaceAreaForReviewPath(file.normalizedPath)),
-  ).size;
-  const contractSurfaceChanged = target.tags.includes('frontend_contract') ||
-    target.tags.includes('desktop_contract') ||
-    target.tags.includes('web_server_contract') ||
-    target.tags.includes('api_layer') ||
-    target.tags.includes('transport');
-  const totalLinesChanged = changeStats.totalLinesChanged;
-  const factors: ReviewTeamRiskFactors = {
-    fileCount: changeStats.fileCount,
-    ...(totalLinesChanged !== undefined ? { totalLinesChanged } : {}),
-    lineCountSource: changeStats.lineCountSource,
-    securityFileCount,
-    workspaceAreaCount,
-    contractSurfaceChanged,
-  };
-
-  if (target.resolution === 'unknown' || changeStats.fileCount === 0) {
-    return {
-      strategyLevel: 'normal',
-      score: 0,
-      rationale: 'unresolved target; keep a conservative normal review recommendation.',
-      factors,
-    };
-  }
-
-  const lineScore =
-    totalLinesChanged === undefined
-      ? 0
-      : Math.floor(totalLinesChanged / 100);
-  const crossAreaScore = Math.max(0, workspaceAreaCount - 1) * 2;
-  const score =
-    changeStats.fileCount +
-    lineScore +
-    securityFileCount * 3 +
-    crossAreaScore +
-    (contractSurfaceChanged ? 2 : 0);
-  const strategyLevel: ReviewStrategyLevel =
-    score <= 5
-      ? 'quick'
-      : score <= 20
-        ? 'normal'
-        : 'deep';
-  const sizeLabel = totalLinesChanged === undefined
-    ? `${changeStats.fileCount} files, unknown lines`
-    : `${changeStats.fileCount} files, ${totalLinesChanged} lines`;
-  const riskDetails = [
-    pluralize(securityFileCount, 'security-sensitive file'),
-    pluralize(workspaceAreaCount, 'workspace area'),
-    contractSurfaceChanged ? 'contract surface changed' : undefined,
-  ].filter(Boolean).join(', ');
-  const rationale =
-    strategyLevel === 'quick'
-      ? `Small change (${sizeLabel}). Quick scan sufficient.`
-      : strategyLevel === 'normal'
-        ? `Medium change (${sizeLabel}; ${riskDetails}). Standard review recommended.`
-        : `Large/high-risk change (${sizeLabel}; ${riskDetails}). Deep review recommended.`;
-
-  return {
-    strategyLevel,
-    score,
-    rationale,
-    factors,
-  };
-}
-
-const REVIEW_STRATEGY_RANK: Record<ReviewStrategyLevel, number> = {
-  quick: 0,
-  normal: 1,
-  deep: 2,
-};
-
-function crossCrateChangeCountForReviewTarget(
-  target: ReviewTargetClassification,
-): number {
-  const crateNames = new Set(
-    target.files
-      .filter((file) => !file.excluded)
-      .map((file) => /^src\/crates\/([^/]+)/.exec(file.normalizedPath)?.[1])
-      .filter((crateName): crateName is string => Boolean(crateName)),
-  );
-
-  return Math.max(0, crateNames.size - 1);
-}
-
-function buildBackendCompatibleRiskFactors(
-  target: ReviewTargetClassification,
-  changeStats: ReviewTeamChangeStats,
-): ReviewTeamBackendRiskFactors {
-  const includedFiles = target.files.filter((file) => !file.excluded);
-
-  return {
-    fileCount: changeStats.fileCount,
-    totalLinesChanged: changeStats.totalLinesChanged ?? 0,
-    lineCountSource: changeStats.lineCountSource,
-    filesInSecurityPaths: includedFiles.filter((file) =>
-      isSecuritySensitiveReviewPath(file.normalizedPath),
-    ).length,
-    crossCrateChanges: crossCrateChangeCountForReviewTarget(target),
-    maxCyclomaticComplexityDelta: 0,
-    maxCyclomaticComplexityDeltaSource: 'not_measured',
-  };
-}
-
-function recommendBackendCompatibleStrategyForTarget(
-  target: ReviewTargetClassification,
-  changeStats: ReviewTeamChangeStats,
-): ReviewTeamBackendStrategyRecommendation {
-  const factors = buildBackendCompatibleRiskFactors(target, changeStats);
-  const score =
-    factors.fileCount +
-    Math.floor(factors.totalLinesChanged / 100) +
-    factors.filesInSecurityPaths * 3 +
-    factors.crossCrateChanges * 2;
-  const strategyLevel: ReviewStrategyLevel =
-    score <= 5
-      ? 'quick'
-      : score <= 20
-        ? 'normal'
-        : 'deep';
-  const rationale =
-    strategyLevel === 'quick'
-      ? `Backend-compatible policy sees a small change (${factors.fileCount} files, ${factors.totalLinesChanged} lines).`
-      : strategyLevel === 'normal'
-        ? `Backend-compatible policy sees a medium change (${factors.fileCount} files, ${factors.totalLinesChanged} lines).`
-        : `Backend-compatible policy sees a large/high-risk change (${factors.fileCount} files, ${factors.totalLinesChanged} lines, ${factors.filesInSecurityPaths} security files).`;
-
-  return {
-    strategyLevel,
-    score,
-    rationale,
-    factors,
-  };
-}
-
-function resolveStrategyMismatchSeverity(params: {
-  finalStrategy: ReviewStrategyLevel;
-  frontendRecommendation: ReviewStrategyLevel;
-  backendRecommendation: ReviewStrategyLevel;
-}): ReviewTeamStrategyMismatchSeverity {
-  const finalRank = REVIEW_STRATEGY_RANK[params.finalStrategy];
-  const recommendedRank = Math.max(
-    REVIEW_STRATEGY_RANK[params.frontendRecommendation],
-    REVIEW_STRATEGY_RANK[params.backendRecommendation],
-  );
-  const distance = Math.abs(finalRank - recommendedRank);
-
-  if (distance === 0) {
-    return 'none';
-  }
-  if (distance >= 2) {
-    return 'high';
-  }
-  return finalRank < recommendedRank ? 'medium' : 'low';
-}
-
-function buildReviewStrategyDecision(params: {
-  teamDefaultStrategy: ReviewStrategyLevel;
-  finalStrategy: ReviewStrategyLevel;
-  userOverride?: ReviewStrategyLevel;
-  frontendRecommendation: ReviewTeamStrategyRecommendation;
-  backendRecommendation: ReviewTeamBackendStrategyRecommendation;
-}): ReviewTeamStrategyDecision {
-  const mismatch =
-    params.finalStrategy !== params.frontendRecommendation.strategyLevel ||
-    params.finalStrategy !== params.backendRecommendation.strategyLevel;
-  const mismatchSeverity = resolveStrategyMismatchSeverity({
-    finalStrategy: params.finalStrategy,
-    frontendRecommendation: params.frontendRecommendation.strategyLevel,
-    backendRecommendation: params.backendRecommendation.strategyLevel,
-  });
-  const recommendationSummary = [
-    `frontend=${params.frontendRecommendation.strategyLevel}`,
-    `backend=${params.backendRecommendation.strategyLevel}`,
-  ].join(', ');
-
-  return {
-    authority: 'mismatch_warning',
-    teamDefaultStrategy: params.teamDefaultStrategy,
-    ...(params.userOverride ? { userOverride: params.userOverride } : {}),
-    finalStrategy: params.finalStrategy,
-    frontendRecommendation: params.frontendRecommendation,
-    backendRecommendation: params.backendRecommendation,
-    mismatch,
-    mismatchSeverity,
-    rationale: mismatch
-      ? `Final strategy ${params.finalStrategy} differs from advisory recommendations (${recommendationSummary}); keep this as non-blocking launch/report metadata.`
-      : `Final strategy ${params.finalStrategy} matches advisory recommendations (${recommendationSummary}).`,
-  };
-}
-
-function buildWorkPacketScopeFromFiles(
-  target: ReviewTargetClassification,
-  files: string[],
-  group?: { index: number; count: number },
-): ReviewTeamWorkPacketScope {
-  return {
-    kind: 'review_target',
-    targetSource: target.source,
-    targetResolution: target.resolution,
-    targetTags: [...target.tags],
-    fileCount: files.length,
-    files,
-    excludedFileCount:
-      target.files.length - target.files.filter((file) => !file.excluded).length,
-    ...(group ? { groupIndex: group.index, groupCount: group.count } : {}),
-  };
-}
-
-function buildWorkPacket(params: {
-  member: ReviewTeamMember;
-  phase: ReviewTeamWorkPacket['phase'];
-  launchBatch: number;
-  scope: ReviewTeamWorkPacketScope;
-  timeoutSeconds: number;
-}): ReviewTeamWorkPacket {
-  const manifestMember = toManifestMember(params.member);
-  const packetGroupSuffix =
-    params.phase === 'reviewer' &&
-    params.scope.groupIndex !== undefined &&
-    params.scope.groupCount !== undefined
-      ? `:group-${params.scope.groupIndex}-of-${params.scope.groupCount}`
-      : '';
-
-  return {
-    packetId: `${params.phase}:${manifestMember.subagentId}${packetGroupSuffix}`,
-    phase: params.phase,
-    launchBatch: params.launchBatch,
-    subagentId: manifestMember.subagentId,
-    displayName: manifestMember.displayName,
-    roleName: manifestMember.roleName,
-    assignedScope: params.scope,
-    allowedTools: [...params.member.allowedTools],
-    timeoutSeconds: params.timeoutSeconds,
-    requiredOutputFields:
-      params.phase === 'judge'
-        ? [...JUDGE_WORK_PACKET_REQUIRED_OUTPUT_FIELDS]
-        : [...REVIEWER_WORK_PACKET_REQUIRED_OUTPUT_FIELDS],
-    strategyLevel: manifestMember.strategyLevel,
-    strategyDirective: manifestMember.strategyDirective,
-    model: manifestMember.model || DEFAULT_REVIEW_TEAM_MODEL,
-  };
-}
-
-function splitFilesIntoGroups(files: string[], groupCount: number): string[][] {
-  if (groupCount <= 1) {
-    return [files];
-  }
-
-  const groups: string[][] = [];
-  let cursor = 0;
-  for (let index = 0; index < groupCount; index += 1) {
-    const remainingFiles = files.length - cursor;
-    const remainingGroups = groupCount - index;
-    const groupSize = Math.ceil(remainingFiles / remainingGroups);
-    groups.push(files.slice(cursor, cursor + groupSize));
-    cursor += groupSize;
-  }
-  return groups;
-}
-
-interface WorkspaceAreaFileBucket {
-  key: string;
-  index: number;
-  files: string[];
-}
-
-function groupFilesByWorkspaceArea(files: string[]): WorkspaceAreaFileBucket[] {
-  const buckets: WorkspaceAreaFileBucket[] = [];
-  const bucketByKey = new Map<string, WorkspaceAreaFileBucket>();
-
-  for (const file of files) {
-    const key = workspaceAreaForReviewPath(file);
-    let bucket = bucketByKey.get(key);
-    if (!bucket) {
-      bucket = {
-        key,
-        index: buckets.length,
-        files: [],
-      };
-      buckets.push(bucket);
-      bucketByKey.set(key, bucket);
-    }
-    bucket.files.push(file);
-  }
-
-  return buckets;
-}
-
-function splitFilesIntoModuleAwareGroups(
-  files: string[],
-  groupCount: number,
-): string[][] {
-  if (groupCount <= 1) {
-    return [files];
-  }
-
-  const buckets = groupFilesByWorkspaceArea(files);
-  if (buckets.length <= 1) {
-    return splitFilesIntoGroups(files, groupCount);
-  }
-
-  if (buckets.length >= groupCount) {
-    const groups = Array.from({ length: groupCount }, () => [] as string[]);
-    const sortedBuckets = [...buckets].sort(
-      (a, b) => b.files.length - a.files.length || a.index - b.index,
-    );
-
-    for (const bucket of sortedBuckets) {
-      let targetIndex = 0;
-      for (let index = 1; index < groups.length; index += 1) {
-        if (groups[index].length < groups[targetIndex].length) {
-          targetIndex = index;
-        }
-      }
-      groups[targetIndex].push(...bucket.files);
-    }
-
-    return groups.filter((group) => group.length > 0);
-  }
-
-  const chunkCounts = buckets.map(() => 1);
-  let remainingChunks = groupCount - buckets.length;
-  while (remainingChunks > 0) {
-    let targetBucketIndex = -1;
-    let largestAverageChunkSize = 0;
-
-    for (let index = 0; index < buckets.length; index += 1) {
-      if (chunkCounts[index] >= buckets[index].files.length) {
-        continue;
-      }
-      const averageChunkSize = buckets[index].files.length / chunkCounts[index];
-      if (averageChunkSize > largestAverageChunkSize) {
-        largestAverageChunkSize = averageChunkSize;
-        targetBucketIndex = index;
-      }
-    }
-
-    if (targetBucketIndex === -1) {
-      break;
-    }
-
-    chunkCounts[targetBucketIndex] += 1;
-    remainingChunks -= 1;
-  }
-
-  return buckets.flatMap((bucket, index) =>
-    splitFilesIntoGroups(bucket.files, chunkCounts[index]),
-  );
-}
-
-function effectiveMaxSameRoleInstances(params: {
-  executionPolicy: ReviewTeamExecutionPolicy;
-  concurrencyPolicy: ReviewTeamConcurrencyPolicy;
-  reviewerMemberCount: number;
-}): number {
-  const reviewerMemberCount = Math.max(1, params.reviewerMemberCount);
-  const maxPerRole = Math.floor(
-    params.concurrencyPolicy.maxParallelInstances / reviewerMemberCount,
-  );
-
-  return Math.max(
-    1,
-    Math.min(params.executionPolicy.maxSameRoleInstances, Math.max(1, maxPerRole)),
-  );
-}
-
-function resolveReviewerPacketScopes(
-  target: ReviewTargetClassification,
-  executionPolicy: ReviewTeamExecutionPolicy,
-  concurrencyPolicy: ReviewTeamConcurrencyPolicy,
-  reviewerMemberCount: number,
-): ReviewTeamWorkPacketScope[] {
-  const includedFiles = target.files
-    .filter((file) => !file.excluded)
-    .map((file) => file.normalizedPath);
-  const shouldSplit =
-    executionPolicy.reviewerFileSplitThreshold > 0 &&
-    executionPolicy.maxSameRoleInstances > 1 &&
-    includedFiles.length > executionPolicy.reviewerFileSplitThreshold;
-
-  if (!shouldSplit) {
-    return [buildWorkPacketScopeFromFiles(target, includedFiles)];
-  }
-
-  const maxSameRoleInstances = effectiveMaxSameRoleInstances({
-    executionPolicy,
-    concurrencyPolicy,
-    reviewerMemberCount,
-  });
-  const groupCount = Math.min(
-    maxSameRoleInstances,
-    Math.ceil(includedFiles.length / executionPolicy.reviewerFileSplitThreshold),
-  );
-  if (groupCount <= 1) {
-    return [buildWorkPacketScopeFromFiles(target, includedFiles)];
-  }
-
-  const fileGroups = splitFilesIntoModuleAwareGroups(includedFiles, groupCount);
-  return fileGroups.map((files, index) =>
-    buildWorkPacketScopeFromFiles(target, files, {
-      index: index + 1,
-      count: fileGroups.length,
-    }),
-  );
-}
-
-function buildWorkPackets(params: {
-  reviewerMembers: ReviewTeamMember[];
-  judgeMember?: ReviewTeamMember;
-  target: ReviewTargetClassification;
-  executionPolicy: ReviewTeamExecutionPolicy;
-  concurrencyPolicy: ReviewTeamConcurrencyPolicy;
-}): ReviewTeamWorkPacket[] {
-  const reviewerScopes = resolveReviewerPacketScopes(
-    params.target,
-    params.executionPolicy,
-    params.concurrencyPolicy,
-    params.reviewerMembers.length,
-  );
-  const fullScope = buildWorkPacketScopeFromFiles(
-    params.target,
-    params.target.files
-      .filter((file) => !file.excluded)
-      .map((file) => file.normalizedPath),
-  );
-  const reviewerSeeds = params.reviewerMembers.flatMap((member) =>
-    reviewerScopes.map((scope) => ({ member, scope })),
-  );
-  const orderedReviewerSeeds = params.concurrencyPolicy.batchExtrasSeparately
-    ? [
-      ...reviewerSeeds.filter((seed) => seed.member.source === 'core'),
-      ...reviewerSeeds.filter((seed) => seed.member.source === 'extra'),
-    ]
-    : reviewerSeeds;
-  const reviewerPackets = orderedReviewerSeeds.map((seed, index) =>
-    buildWorkPacket({
-      member: seed.member,
-      phase: 'reviewer',
-      launchBatch:
-        Math.floor(index / params.concurrencyPolicy.maxParallelInstances) + 1,
-      scope: seed.scope,
-      timeoutSeconds: params.executionPolicy.reviewerTimeoutSeconds,
-    }),
-  );
-  const finalReviewerBatch = reviewerPackets.reduce(
-    (maxBatch, packet) => Math.max(maxBatch, packet.launchBatch),
-    0,
-  );
-  const judgePacket = params.judgeMember
-    ? [
-      buildWorkPacket({
-        member: params.judgeMember,
-        phase: 'judge',
-        launchBatch: finalReviewerBatch + 1,
-        scope: fullScope,
-        timeoutSeconds: params.executionPolicy.judgeTimeoutSeconds,
-      }),
-    ]
-    : [];
-
-  return [...reviewerPackets, ...judgePacket];
-}
-
-const SHARED_CONTEXT_CACHE_ENTRY_LIMIT = 80;
-const SHARED_CONTEXT_CACHE_RECOMMENDED_TOOLS: ReviewTeamSharedContextTool[] = [
-  'GetFileDiff',
-  'Read',
-];
-
-function buildSharedContextCachePlan(
-  workPackets: ReviewTeamWorkPacket[] = [],
-): ReviewTeamSharedContextCachePlan {
-  const fileContextByPath = new Map<
-    string,
-    {
-      path: string;
-      workspaceArea: string;
-      consumerPacketIds: string[];
-      firstSeenIndex: number;
-    }
-  >();
-  let nextSeenIndex = 0;
-
-  for (const packet of workPackets) {
-    if (packet.phase !== 'reviewer') {
-      continue;
-    }
-
-    for (const path of packet.assignedScope.files) {
-      let entry = fileContextByPath.get(path);
-      if (!entry) {
-        entry = {
-          path,
-          workspaceArea: workspaceAreaForReviewPath(path),
-          consumerPacketIds: [],
-          firstSeenIndex: nextSeenIndex,
-        };
-        nextSeenIndex += 1;
-        fileContextByPath.set(path, entry);
-      }
-      if (!entry.consumerPacketIds.includes(packet.packetId)) {
-        entry.consumerPacketIds.push(packet.packetId);
-      }
-    }
-  }
-
-  const repeatedFileContexts = Array.from(fileContextByPath.values())
-    .filter((entry) => entry.consumerPacketIds.length > 1)
-    .sort((a, b) => a.firstSeenIndex - b.firstSeenIndex);
-  const entries = repeatedFileContexts
-    .slice(0, SHARED_CONTEXT_CACHE_ENTRY_LIMIT)
-    .map((entry, index) => ({
-      cacheKey: `shared-context:${index + 1}`,
-      path: entry.path,
-      workspaceArea: entry.workspaceArea,
-      recommendedTools: [...SHARED_CONTEXT_CACHE_RECOMMENDED_TOOLS],
-      consumerPacketIds: entry.consumerPacketIds,
-    }));
-
-  return {
-    source: 'work_packets',
-    strategy: 'reuse_readonly_file_context_by_cache_key',
-    entries,
-    omittedEntryCount: Math.max(
-      0,
-      repeatedFileContexts.length - SHARED_CONTEXT_CACHE_ENTRY_LIMIT,
-    ),
-  };
-}
-
-const INCREMENTAL_REVIEW_CACHE_INVALIDATIONS: ReviewTeamIncrementalReviewCacheInvalidation[] = [
-  'target_file_set_changed',
-  'target_line_count_changed',
-  'target_tag_changed',
-  'target_warning_changed',
-  'reviewer_roster_changed',
-  'strategy_changed',
-];
-
-function stableFingerprint(input: unknown): string {
-  const serialized = JSON.stringify(input);
-  let hash = 0x811c9dc5;
-  for (let index = 0; index < serialized.length; index += 1) {
-    hash ^= serialized.charCodeAt(index);
-    hash = Math.imul(hash, 0x01000193);
-  }
-  return (hash >>> 0).toString(16).padStart(8, '0');
-}
-
-function buildIncrementalReviewCachePlan(params: {
-  target: ReviewTargetClassification;
-  changeStats: ReviewTeamChangeStats;
-  strategyLevel: ReviewStrategyLevel;
-  workPackets: ReviewTeamWorkPacket[];
-}): ReviewTeamIncrementalReviewCachePlan {
-  const filePaths = params.target.files
-    .filter((file) => !file.excluded)
-    .map((file) => file.normalizedPath)
-    .sort((a, b) => a.localeCompare(b));
-  const workspaceAreas = Array.from(
-    new Set(filePaths.map((file) => workspaceAreaForReviewPath(file))),
-  ).sort((a, b) => a.localeCompare(b));
-  const targetTags = [...params.target.tags].sort((a, b) => a.localeCompare(b));
-  const targetWarnings = params.target.warnings
-    .map((warning) => warning.code)
-    .sort((a, b) => a.localeCompare(b));
-  const reviewerPacketIds = params.workPackets
-    .filter((packet) => packet.phase === 'reviewer')
-    .map((packet) => packet.packetId)
-    .sort((a, b) => a.localeCompare(b));
-  const fingerprint = stableFingerprint({
-    source: params.target.source,
-    resolution: params.target.resolution,
-    filePaths,
-    workspaceAreas,
-    targetTags,
-    targetWarnings,
-    lineCount: params.changeStats.totalLinesChanged ?? null,
-    lineCountSource: params.changeStats.lineCountSource,
-    reviewerPacketIds,
-    strategyLevel: params.strategyLevel,
-  });
-
-  return {
-    source: 'target_manifest',
-    strategy: 'reuse_completed_packets_when_fingerprint_matches',
-    cacheKey: `incremental-review:${fingerprint}`,
-    fingerprint,
-    filePaths,
-    workspaceAreas,
-    targetTags,
-    reviewerPacketIds,
-    ...(params.changeStats.totalLinesChanged !== undefined
-      ? { lineCount: params.changeStats.totalLinesChanged }
-      : {}),
-    lineCountSource: params.changeStats.lineCountSource,
-    invalidatesOn: [...INCREMENTAL_REVIEW_CACHE_INVALIDATIONS],
-  };
-}
-
-function predictTimeoutSeconds(params: {
-  role: 'reviewer' | 'judge';
-  strategyLevel: ReviewStrategyLevel;
-  changeStats: ReviewTeamChangeStats;
-  reviewerCount: number;
-}): number {
-  const totalLinesChanged = params.changeStats.totalLinesChanged ?? 0;
-  const base = PREDICTIVE_TIMEOUT_BASE_SECONDS[params.strategyLevel];
-  const raw =
-    base +
-    params.changeStats.fileCount * PREDICTIVE_TIMEOUT_PER_FILE_SECONDS +
-    Math.floor(totalLinesChanged / 100) *
-      PREDICTIVE_TIMEOUT_PER_100_LINES_SECONDS;
-  const reviewerCount = Math.max(1, params.reviewerCount);
-  const multiplier =
-    params.role === 'judge'
-      ? 1 + Math.floor((reviewerCount - 1) / 3)
-      : 1;
-
-  return Math.min(raw * multiplier, MAX_PREDICTIVE_TIMEOUT_SECONDS);
-}
-
-function buildEffectiveExecutionPolicy(params: {
-  basePolicy: ReviewTeamExecutionPolicy;
-  strategyLevel: ReviewStrategyLevel;
-  target: ReviewTargetClassification;
-  changeStats: ReviewTeamChangeStats;
-  reviewerCount: number;
-}): ReviewTeamExecutionPolicy {
-  if (
-    params.target.resolution === 'unknown' &&
-    params.changeStats.fileCount === 0 &&
-    params.changeStats.totalLinesChanged === undefined
-  ) {
-    return params.basePolicy;
-  }
-
-  const reviewerTimeoutSeconds = predictTimeoutSeconds({
-    role: 'reviewer',
-    strategyLevel: params.strategyLevel,
-    changeStats: params.changeStats,
-    reviewerCount: params.reviewerCount,
-  });
-  const judgeTimeoutSeconds = predictTimeoutSeconds({
-    role: 'judge',
-    strategyLevel: params.strategyLevel,
-    changeStats: params.changeStats,
-    reviewerCount: params.reviewerCount,
-  });
-
-  return {
-    ...params.basePolicy,
-    reviewerTimeoutSeconds:
-      params.basePolicy.reviewerTimeoutSeconds === 0
-        ? 0
-        : Math.max(
-          params.basePolicy.reviewerTimeoutSeconds,
-          reviewerTimeoutSeconds,
-        ),
-    judgeTimeoutSeconds:
-      params.basePolicy.judgeTimeoutSeconds === 0
-        ? 0
-        : Math.max(
-          params.basePolicy.judgeTimeoutSeconds,
-          judgeTimeoutSeconds,
-        ),
-  };
-}
-
-function estimateChangedLinesForScope(params: {
-  scope: ReviewTeamWorkPacketScope;
-  changeStats: ReviewTeamChangeStats;
-  totalIncludedFileCount: number;
-}): number {
-  if (params.changeStats.totalLinesChanged === undefined) {
-    return params.scope.fileCount * PROMPT_BYTE_ESTIMATE_UNKNOWN_LINES_PER_FILE;
-  }
-
-  if (params.totalIncludedFileCount <= 0) {
-    return params.changeStats.totalLinesChanged;
-  }
-
-  return Math.ceil(
-    params.changeStats.totalLinesChanged *
-      (params.scope.fileCount / params.totalIncludedFileCount),
-  );
-}
-
-function estimateReviewerPromptBytes(params: {
-  packet: ReviewTeamWorkPacket;
-  changeStats: ReviewTeamChangeStats;
-  totalIncludedFileCount: number;
-}): number {
-  const pathBytes = params.packet.assignedScope.files.reduce(
-    (total, filePath) => total + filePath.length + 1,
-    0,
-  );
-  const estimatedChangedLines = estimateChangedLinesForScope({
-    scope: params.packet.assignedScope,
-    changeStats: params.changeStats,
-    totalIncludedFileCount: params.totalIncludedFileCount,
-  });
-
-  return Math.ceil(
-    PROMPT_BYTE_ESTIMATE_BASE_BYTES +
-      pathBytes +
-      params.packet.assignedScope.fileCount * PROMPT_BYTE_ESTIMATE_PER_FILE_BYTES +
-      estimatedChangedLines * PROMPT_BYTE_ESTIMATE_PER_CHANGED_LINE_BYTES,
-  );
-}
-
-function estimateMaxReviewerPromptBytes(params: {
-  workPackets: ReviewTeamWorkPacket[];
-  target: ReviewTargetClassification;
-  changeStats: ReviewTeamChangeStats;
-}): number {
-  const reviewerPackets = params.workPackets.filter(
-    (packet) => packet.phase === 'reviewer',
-  );
-  const totalIncludedFileCount = params.target.files.filter(
-    (file) => !file.excluded,
-  ).length;
-
-  if (reviewerPackets.length === 0) {
-    return PROMPT_BYTE_ESTIMATE_BASE_BYTES;
-  }
-
-  return Math.max(
-    ...reviewerPackets.map((packet) =>
-      estimateReviewerPromptBytes({
-        packet,
-        changeStats: params.changeStats,
-        totalIncludedFileCount,
-      }),
-    ),
-  );
-}
-
-function buildTokenBudgetPlan(params: {
-  mode: ReviewTokenBudgetMode;
-  activeReviewerCalls: number;
-  eligibleExtraReviewerCount: number;
-  maxExtraReviewers: number;
-  skippedReviewerIds: string[];
-  target: ReviewTargetClassification;
-  changeStats: ReviewTeamChangeStats;
-  executionPolicy: ReviewTeamExecutionPolicy;
-  workPackets: ReviewTeamWorkPacket[];
-}): ReviewTeamTokenBudgetPlan {
-  const includedFileCount = params.target.files.filter(
-    (file) => !file.excluded,
-  ).length;
-  const fileSplitGuardrailActive =
-    params.executionPolicy.reviewerFileSplitThreshold > 0 &&
-    includedFileCount > params.executionPolicy.reviewerFileSplitThreshold;
-  const maxPromptBytesPerReviewer =
-    TOKEN_BUDGET_PROMPT_BYTE_LIMIT_BY_MODE[params.mode];
-  const estimatedPromptBytesPerReviewer = estimateMaxReviewerPromptBytes({
-    workPackets: params.workPackets,
-    target: params.target,
-    changeStats: params.changeStats,
-  });
-  const promptByteLimitExceeded =
-    estimatedPromptBytesPerReviewer > maxPromptBytesPerReviewer;
-  const largeDiffSummaryFirst = promptByteLimitExceeded;
-  const decisions: ReviewTeamTokenBudgetDecision[] = [];
-  const warnings: string[] = [];
-
-  if (promptByteLimitExceeded) {
-    decisions.push({
-      kind: 'summary_first_full_scope',
-      reason: 'prompt_bytes_exceeded',
-      detail:
-        `Estimated reviewer prompt ${estimatedPromptBytesPerReviewer} bytes exceeds ${maxPromptBytesPerReviewer} bytes for ${params.mode} budget; use summary-first while keeping every assigned_scope file visible.`,
-    });
-    warnings.push(
-      'Estimated reviewer prompt exceeds the selected token budget; use summary-first without hiding assigned files.',
-    );
-  }
-
-  if (params.skippedReviewerIds.length > 0) {
-    decisions.push({
-      kind: 'skip_extra_reviewers',
-      reason: 'extra_reviewers_skipped',
-      detail:
-        'Some extra reviewers were skipped by the selected token budget mode.',
-      affectedReviewerIds: [...params.skippedReviewerIds],
-    });
-    warnings.push(
-      'Some extra reviewers were skipped by the selected token budget mode.',
-    );
-  }
-
-  return {
-    mode: params.mode,
-    estimatedReviewerCalls: params.activeReviewerCalls,
-    maxReviewerCalls:
-      params.activeReviewerCalls +
-      Math.max(0, params.eligibleExtraReviewerCount - params.maxExtraReviewers),
-    maxExtraReviewers: params.maxExtraReviewers,
-    ...(fileSplitGuardrailActive
-      ? { maxFilesPerReviewer: params.executionPolicy.reviewerFileSplitThreshold }
-      : {}),
-    maxPromptBytesPerReviewer,
-    estimatedPromptBytesPerReviewer,
-    promptByteEstimateSource: 'manifest_heuristic',
-    promptByteLimitExceeded,
-    largeDiffSummaryFirst,
-    decisions,
-    skippedReviewerIds: params.skippedReviewerIds,
-    warnings,
-  };
 }
 
 export function buildEffectiveReviewTeamManifest(
@@ -2283,6 +1259,7 @@ export function buildEffectiveReviewTeamManifest(
     options.rateLimitStatus,
   );
   const strategyLevel = options.strategyOverride ?? team.strategyLevel;
+  const scopeProfile = buildDeepReviewScopeProfile(strategyLevel);
   const strategyRecommendation = recommendReviewStrategyForTarget(target, changeStats);
   const backendStrategyRecommendation = recommendBackendCompatibleStrategyForTarget(
     target,
@@ -2344,6 +1321,12 @@ export function buildEffectiveReviewTeamManifest(
     executionPolicy,
     concurrencyPolicy,
   });
+  const evidencePack = buildDeepReviewEvidencePack({
+    target,
+    changeStats,
+    scopeProfile,
+    workPackets,
+  });
   const sharedContextCache = buildSharedContextCachePlan(workPackets);
   const incrementalReviewCache = buildIncrementalReviewCachePlan({
     target,
@@ -2388,12 +1371,14 @@ export function buildEffectiveReviewTeamManifest(
     policySource: options.policySource ?? 'default-review-team-config',
     target,
     strategyLevel,
+    scopeProfile,
     strategyRecommendation,
     strategyDecision,
     executionPolicy,
     concurrencyPolicy,
     changeStats,
     preReviewSummary,
+    evidencePack,
     sharedContextCache,
     incrementalReviewCache,
     tokenBudget,
@@ -2405,332 +1390,9 @@ export function buildEffectiveReviewTeamManifest(
   };
 }
 
-function formatResponsibilities(items: string[]): string {
-  return items.map((item) => `    - ${item}`).join('\n');
-}
-
-function formatStrategyImpact(
-  strategyLevel: ReviewStrategyLevel,
-  strategyProfiles: Record<ReviewStrategyLevel, ReviewStrategyProfile> = REVIEW_STRATEGY_PROFILES,
-): string {
-  const definition = strategyProfiles[strategyLevel];
-  return `Token/time impact: approximately ${definition.tokenImpact} token usage and ${definition.runtimeImpact} runtime.`;
-}
-
-function formatManifestList(
-  members: ReviewTeamManifestMember[],
-  emptyValue: string,
-): string {
-  if (members.length === 0) {
-    return emptyValue;
-  }
-
-  return members
-    .map((member) =>
-      member.reason
-        ? `${member.subagentId}: ${member.reason}`
-        : member.subagentId,
-    )
-    .join(', ');
-}
-
-function workPacketToPromptPayload(packet: ReviewTeamWorkPacket) {
-  return {
-    packet_id: packet.packetId,
-    phase: packet.phase,
-    launch_batch: packet.launchBatch,
-    subagent_type: packet.subagentId,
-    display_name: packet.displayName,
-    role: packet.roleName,
-    assigned_scope: {
-      kind: packet.assignedScope.kind,
-      target_source: packet.assignedScope.targetSource,
-      target_resolution: packet.assignedScope.targetResolution,
-      target_tags: packet.assignedScope.targetTags,
-      file_count: packet.assignedScope.fileCount,
-      files: packet.assignedScope.files,
-      excluded_file_count: packet.assignedScope.excludedFileCount,
-      ...(packet.assignedScope.groupIndex !== undefined
-        ? { group_index: packet.assignedScope.groupIndex }
-        : {}),
-      ...(packet.assignedScope.groupCount !== undefined
-        ? { group_count: packet.assignedScope.groupCount }
-        : {}),
-    },
-    allowed_tools: packet.allowedTools,
-    timeout_seconds: packet.timeoutSeconds,
-    required_output_fields: packet.requiredOutputFields,
-    strategy: packet.strategyLevel,
-    model_id: packet.model,
-    prompt_directive: packet.strategyDirective,
-  };
-}
-
-function formatWorkPacketBlock(workPackets: ReviewTeamWorkPacket[] = []): string {
-  if (workPackets.length === 0) {
-    return '- none';
-  }
-
-  return [
-    '```json',
-    JSON.stringify(workPackets.map(workPacketToPromptPayload), null, 2),
-    '```',
-  ].join('\n');
-}
-
-function formatPreReviewSummaryBlock(summary: ReviewTeamPreReviewSummary): string {
-  return [
-    'Pre-generated diff summary:',
-    '```json',
-    JSON.stringify(summary, null, 2),
-    '```',
-  ].join('\n');
-}
-
-function sharedContextCacheToPromptPayload(plan: ReviewTeamSharedContextCachePlan) {
-  return {
-    source: plan.source,
-    strategy: plan.strategy,
-    omitted_entry_count: plan.omittedEntryCount,
-    entries: plan.entries.map((entry) => ({
-      cache_key: entry.cacheKey,
-      path: entry.path,
-      workspace_area: entry.workspaceArea,
-      recommended_tools: entry.recommendedTools,
-      consumer_packet_ids: entry.consumerPacketIds,
-    })),
-  };
-}
-
-function formatSharedContextCacheBlock(plan: ReviewTeamSharedContextCachePlan): string {
-  return [
-    'Shared context cache plan:',
-    '```json',
-    JSON.stringify(sharedContextCacheToPromptPayload(plan), null, 2),
-    '```',
-  ].join('\n');
-}
-
-function incrementalReviewCacheToPromptPayload(plan: ReviewTeamIncrementalReviewCachePlan) {
-  return {
-    source: plan.source,
-    strategy: plan.strategy,
-    cache_key: plan.cacheKey,
-    fingerprint: plan.fingerprint,
-    file_paths: plan.filePaths,
-    workspace_areas: plan.workspaceAreas,
-    target_tags: plan.targetTags,
-    reviewer_packet_ids: plan.reviewerPacketIds,
-    ...(plan.lineCount !== undefined ? { line_count: plan.lineCount } : {}),
-    line_count_source: plan.lineCountSource,
-    invalidates_on: plan.invalidatesOn,
-  };
-}
-
-function formatIncrementalReviewCacheBlock(plan: ReviewTeamIncrementalReviewCachePlan): string {
-  return [
-    'Incremental review cache plan:',
-    '```json',
-    JSON.stringify(incrementalReviewCacheToPromptPayload(plan), null, 2),
-    '```',
-  ].join('\n');
-}
-
-function formatTokenBudgetDecisionKinds(
-  decisions: ReviewTeamTokenBudgetDecision[] = [],
-): string {
-  return decisions.length > 0
-    ? decisions.map((decision) => decision.kind).join(', ')
-    : 'none';
-}
-
 export function buildReviewTeamPromptBlock(
   team: ReviewTeam,
   manifest = buildEffectiveReviewTeamManifest(team),
 ): string {
-  const activeSubagentIds = new Set([
-    ...manifest.coreReviewers.map((member) => member.subagentId),
-    ...manifest.enabledExtraReviewers.map((member) => member.subagentId),
-    ...(manifest.qualityGateReviewer
-      ? [manifest.qualityGateReviewer.subagentId]
-      : []),
-  ]);
-  const activeManifestMembers = [
-    ...manifest.coreReviewers,
-    ...(manifest.qualityGateReviewer ? [manifest.qualityGateReviewer] : []),
-    ...manifest.enabledExtraReviewers,
-  ];
-  const manifestMemberBySubagentId = new Map(
-    activeManifestMembers.map((member) => [member.subagentId, member]),
-  );
-  const members = team.members
-    .filter((member) => member.available && activeSubagentIds.has(member.subagentId))
-    .map((member) => {
-      const manifestMember =
-        manifestMemberBySubagentId.get(member.subagentId) ?? toManifestMember(member);
-      return [
-        `- ${manifestMember.displayName}`,
-        `  - subagent_type: ${manifestMember.subagentId}`,
-        `  - preferred_task_label: ${manifestMember.displayName}`,
-        `  - role: ${manifestMember.roleName}`,
-        `  - locked_core_role: ${manifestMember.locked ? 'yes' : 'no'}`,
-        `  - strategy: ${manifestMember.strategyLevel}`,
-        `  - strategy_source: ${manifestMember.strategySource}`,
-        `  - default_model_slot: ${manifestMember.defaultModelSlot}`,
-        `  - model: ${manifestMember.model || DEFAULT_REVIEW_TEAM_MODEL}`,
-        `  - model_id: ${manifestMember.model || DEFAULT_REVIEW_TEAM_MODEL}`,
-        `  - configured_model: ${manifestMember.configuredModel || manifestMember.model || DEFAULT_REVIEW_TEAM_MODEL}`,
-        ...(manifestMember.modelFallbackReason
-          ? [`  - model_fallback: ${manifestMember.modelFallbackReason}`]
-          : []),
-        `  - prompt_directive: ${manifestMember.strategyDirective}`,
-        '  - responsibilities:',
-        formatResponsibilities(member.responsibilities),
-      ].join('\n');
-    })
-    .join('\n');
-  const executionPolicy = [
-    `- reviewer_timeout_seconds: ${manifest.executionPolicy.reviewerTimeoutSeconds}`,
-    `- judge_timeout_seconds: ${manifest.executionPolicy.judgeTimeoutSeconds}`,
-    `- reviewer_file_split_threshold: ${manifest.executionPolicy.reviewerFileSplitThreshold}`,
-    `- max_same_role_instances: ${manifest.executionPolicy.maxSameRoleInstances}`,
-    `- max_retries_per_role: ${manifest.executionPolicy.maxRetriesPerRole}`,
-  ].join('\n');
-  const concurrencyPolicy = [
-    `- max_parallel_instances: ${manifest.concurrencyPolicy.maxParallelInstances}`,
-    `- stagger_seconds: ${manifest.concurrencyPolicy.staggerSeconds}`,
-    `- max_queue_wait_seconds: ${manifest.concurrencyPolicy.maxQueueWaitSeconds}`,
-    `- batch_extras_separately: ${manifest.concurrencyPolicy.batchExtrasSeparately ? 'yes' : 'no'}`,
-    `- allow_provider_capacity_queue: ${manifest.concurrencyPolicy.allowProviderCapacityQueue ? 'yes' : 'no'}`,
-    `- allow_bounded_auto_retry: ${manifest.concurrencyPolicy.allowBoundedAutoRetry ? 'yes' : 'no'}`,
-    `- auto_retry_elapsed_guard_seconds: ${manifest.concurrencyPolicy.autoRetryElapsedGuardSeconds}`,
-  ].join('\n');
-  const targetLineCount =
-    manifest.changeStats?.totalLinesChanged !== undefined
-      ? `${manifest.changeStats.totalLinesChanged}`
-      : 'unknown';
-  const manifestBlock = [
-    'Run manifest:',
-    `- review_mode: ${manifest.reviewMode}`,
-    `- team_strategy: ${manifest.strategyLevel}`,
-    `- strategy_authority: ${manifest.strategyDecision.authority}`,
-    `- final_strategy: ${manifest.strategyDecision.finalStrategy}`,
-    `- frontend_recommended_strategy: ${manifest.strategyDecision.frontendRecommendation.strategyLevel}`,
-    `- backend_recommended_strategy: ${manifest.strategyDecision.backendRecommendation.strategyLevel}`,
-    `- strategy_user_override: ${manifest.strategyDecision.userOverride ?? 'none'}`,
-    `- strategy_mismatch: ${manifest.strategyDecision.mismatch ? 'yes' : 'no'}`,
-    `- strategy_mismatch_severity: ${manifest.strategyDecision.mismatchSeverity}`,
-    `- max_cyclomatic_complexity_delta: ${manifest.strategyDecision.backendRecommendation.factors.maxCyclomaticComplexityDelta}`,
-    `- max_cyclomatic_complexity_delta_source: ${manifest.strategyDecision.backendRecommendation.factors.maxCyclomaticComplexityDeltaSource}`,
-    ...(manifest.strategyRecommendation
-      ? [
-        `- recommended_strategy: ${manifest.strategyRecommendation.strategyLevel}`,
-        `- strategy_recommendation_score: ${manifest.strategyRecommendation.score}`,
-        `- strategy_recommendation_rationale: ${manifest.strategyRecommendation.rationale}`,
-      ]
-      : []),
-    `- workspace_path: ${manifest.workspacePath || 'inherited from current session'}`,
-    `- policy_source: ${manifest.policySource}`,
-    `- target_source: ${manifest.target.source}`,
-    `- target_resolution: ${manifest.target.resolution}`,
-    `- target_tags: ${manifest.target.tags.join(', ') || 'none'}`,
-    `- target_warnings: ${manifest.target.warnings.map((warning) => warning.code).join(', ') || 'none'}`,
-    `- target_file_count: ${manifest.changeStats?.fileCount ?? manifest.target.files.length}`,
-    `- target_line_count: ${targetLineCount}`,
-    `- target_line_count_source: ${manifest.changeStats?.lineCountSource ?? 'unknown'}`,
-    `- token_budget_mode: ${manifest.tokenBudget.mode}`,
-    `- estimated_reviewer_calls: ${manifest.tokenBudget.estimatedReviewerCalls}`,
-    `- max_prompt_bytes_per_reviewer: ${manifest.tokenBudget.maxPromptBytesPerReviewer ?? 'none'}`,
-    `- estimated_prompt_bytes_per_reviewer: ${manifest.tokenBudget.estimatedPromptBytesPerReviewer ?? 'unknown'}`,
-    `- prompt_byte_estimate_source: ${manifest.tokenBudget.promptByteEstimateSource ?? 'none'}`,
-    `- prompt_byte_limit_exceeded: ${manifest.tokenBudget.promptByteLimitExceeded ? 'yes' : 'no'}`,
-    `- token_budget_decisions: ${formatTokenBudgetDecisionKinds(manifest.tokenBudget.decisions)}`,
-    `- budget_limited_reviewers: ${manifest.tokenBudget.skippedReviewerIds.join(', ') || 'none'}`,
-    `- core_reviewers: ${formatManifestList(manifest.coreReviewers, 'none')}`,
-    `- quality_gate_reviewer: ${manifest.qualityGateReviewer?.subagentId || 'none'}`,
-    `- enabled_extra_reviewers: ${formatManifestList(manifest.enabledExtraReviewers, 'none')}`,
-    '- skipped_reviewers:',
-    ...(manifest.skippedReviewers.length > 0
-      ? manifest.skippedReviewers.map(
-        (member) => `  - ${member.subagentId}: ${member.reason || 'skipped'}`,
-      )
-      : ['  - none']),
-  ].join('\n');
-  const strategyProfiles = team.definition?.strategyProfiles ?? REVIEW_STRATEGY_PROFILES;
-  const strategyRules = REVIEW_STRATEGY_LEVELS.map((level) => {
-    const definition = strategyProfiles[level];
-    const roleEntries = Object.entries(definition.roleDirectives) as [ReviewRoleDirectiveKey, string][];
-    const roleLines = roleEntries.map(
-      ([role, directive]) => `    - ${role}: ${directive}`,
-    );
-    return [
-      `- ${level}: ${definition.summary}`,
-      `  - ${formatStrategyImpact(level, strategyProfiles)}`,
-      `  - Default model slot: ${definition.defaultModelSlot}`,
-      `  - Prompt directive (fallback): ${definition.promptDirective}`,
-      `  - Role-specific directives:`,
-      ...roleLines,
-    ].join('\n');
-  }).join('\n');
-  const commonStrategyRules = REVIEW_STRATEGY_COMMON_RULES.reviewerPromptRules
-    .map((rule) => `- ${rule}`)
-    .join('\n');
-
-  return [
-    manifestBlock,
-    formatPreReviewSummaryBlock(manifest.preReviewSummary),
-    formatSharedContextCacheBlock(manifest.sharedContextCache),
-    formatIncrementalReviewCacheBlock(manifest.incrementalReviewCache),
-    'Review work packets:',
-    formatWorkPacketBlock(manifest.workPackets),
-    'Work packet rules:',
-    '- Each reviewer Task prompt must include the matching work packet verbatim.',
-    '- Include the packet_id in each Task description, for example "Security review [packet reviewer:ReviewSecurity:group-1-of-3]".',
-    '- Each reviewer and judge response must echo packet_id and set status to completed, partial_timeout, timed_out, cancelled_by_user, failed, or skipped.',
-    '- If the reviewer reports packet_id itself, mark reviewers[].packet_status_source as reported in the final submit_code_review payload.',
-    '- If the reviewer omits packet_id but the Task was launched from a packet, infer the packet_id from the Task description or work packet and mark packet_status_source as inferred.',
-    '- If packet_id cannot be reported or inferred, mark packet_status_source as missing and explain the confidence impact in coverage_notes.',
-    '- If a reviewer response is missing packet_id or status, the judge must treat that reviewer output as lower confidence instead of discarding the whole review.',
-    '- Use the pre-generated diff summary for initial orientation and token discipline, but verify claims against assigned files or diffs before reporting findings.',
-    '- When prompt_byte_limit_exceeded is yes, use the pre-generated diff summary before detailed reads. Do not remove files from assigned_scope or hide unreviewed files; if a file cannot be covered, report it in coverage_notes and reliability_signals.',
-    '- Use shared_context_cache entries to reuse read-only GetFileDiff/Read context by cache_key across reviewer packets. Do not duplicate full-file reads when a reusable cached diff or file summary already covers the same path.',
-    '- Use incremental_review_cache only when the target fingerprint matches a prior run; preserve completed reviewer outputs by packet_id and rerun only missing, failed, timed-out, or stale packets. If any invalidates_on condition changed, ignore the cache and explain the fresh review boundary.',
-    '- The assigned_scope is the default scope for that packet; only widen it when a critical cross-file dependency requires it and note the reason in coverage_notes.',
-    'Configured code review team:',
-    members || '- No team members available.',
-    'Execution policy:',
-    executionPolicy,
-    'Concurrency policy:',
-    concurrencyPolicy,
-    'Team execution rules:',
-    '- Run only reviewers listed in core_reviewers and enabled_extra_reviewers.',
-    '- Do not launch skipped_reviewers.',
-    '- If a skipped reviewer has reason not_applicable, mention it in coverage notes without treating it as reduced confidence.',
-    '- If a skipped reviewer has reason budget_limited, mention the budget mode and the coverage tradeoff.',
-    '- If a skipped reviewer has reason invalid_tooling, report it as a configuration issue and do not reduce confidence in the reviewers that did run.',
-    '- If target_resolution is unknown, conditional reviewers may be activated conservatively; report that as coverage context.',
-    `- Run the active core reviewer roles first: ${formatManifestList(manifest.coreReviewers, 'none')}.`,
-    '- Launch reviewer Tasks by launch_batch. Do not launch a later reviewer batch until every reviewer Task in the earlier batch has completed, failed, timed out, or returned partial_timeout.',
-    '- Never launch more reviewer Tasks in one batch than max_parallel_instances. If stagger_seconds is greater than 0, wait that many seconds before starting the next launch_batch.',
-    '- Run ReviewJudge only after the reviewer batch finishes, as the quality-gate pass.',
-    '- If other extra reviewers are configured and enabled, run them in parallel with the locked reviewers whenever possible.',
-    '- When a configured member entry provides model_id, pass model_id with that value to the matching Task call.',
-    '- If reviewer_timeout_seconds is greater than 0, pass timeout_seconds with that value to every reviewer Task call.',
-    '- If judge_timeout_seconds is greater than 0, pass timeout_seconds with that value to the ReviewJudge Task call.',
-    '- If a reviewer Task returns status partial_timeout, treat its output as partial evidence: preserve it in reviewers[].partial_output, mark the reviewer status partial_timeout, and mention the confidence impact in coverage_notes.',
-    '- If a reviewer fails or times out without useful partial output, retry that same reviewer at most max_retries_per_role times: reduce its scope, downgrade strategy by one level when possible, use a shorter timeout, and set retry to true on the retry Task call.',
-    '- In the final submit_code_review payload, populate reliability_signals for context_pressure, compression_preserved, partial_reviewer, and user_decision when those conditions apply. Use severity info/warning/action, count when useful, and source runtime/manifest/report/inferred.',
-    '- If reviewer_file_split_threshold is greater than 0 and the target file count exceeds it, split files across multiple same-role reviewer instances only up to the concurrency-capped max_same_role_instances for this run.',
-    '- Prefer module/workspace-area coherent file groups when splitting reviewer work; avoid mixing unrelated workspace areas in the same packet when the group budget allows it.',
-    '- When file splitting is active, each same-role instance must only review its assigned file group. Label instances in the Task description with both group and packet_id (e.g. "Security review [group 1/3] [packet reviewer:ReviewSecurity:group-1-of-3]").',
-    '- Do not run ReviewFixer during the review pass.',
-    '- Wait for explicit user approval before starting any remediation.',
-    '- The Review Quality Inspector acts as a third-party arbiter: it primarily examines reviewer reports for logical consistency and evidence quality, and only uses code inspection tools for targeted spot-checks when a specific claim needs verification.',
-    'Review strategy rules:',
-    `- Team strategy: ${manifest.strategyLevel}. ${formatStrategyImpact(manifest.strategyLevel, strategyProfiles)}`,
-    '- Risk recommendation is advisory; follow team_strategy, member strategy fields, and work-packet strategy for this run unless the user explicitly changes strategy.',
-    commonStrategyRules,
-    'Review strategy profiles:',
-    strategyRules,
-  ].join('\n');
+  return buildReviewTeamPromptBlockContent(team, manifest);
 }
